backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
face_recognition
numpy
dlib
alembic
pydantic-settings
bcrypt==4.0.1
